Separate code for getting information
Part of #14
